+++ package.json
@@ -45,6 +45,7 @@
     "next-themes": "^0.2.1",
     "react": "^18",
     "react-dom": "^18",
+    "react-fast-compare": "^3.2.2",
     "react-magic-motion": "^1.1.2",
     "react-slick": "^0.30.2",
     "sass": "^1.69.5",
